@@ -1,3 +1,5 @@
+@preconcurrency import Crypto
+
 /*
  * Copyright (c) 2025 Contributors to the Eclipse Foundation.
  * SPDX-License-Identifier: Apache-2.0
@@ -7,7 +9,6 @@ import FoundationEssentials
 #else
 import Foundation
 #endif
-import Crypto
 
 extension Biscuit {
     /// A third party public key that has been used to sign a block
@@ -23,6 +24,17 @@ extension Biscuit {
             self.dataRepresentation = key.dataRepresentation
         }
 
+        /// Construct a ThirdPartyKey from an algorithm and its hex-encoded representation
+        /// Throws: Throws a ValidationError if hexString is not a valid hex-encoded string
+        public init(algorithm: SigningAlgorithm, hexString: String) throws {
+            self.algorithm = algorithm
+            if let data = hexDecode(hexString[...]) {
+                self.dataRepresentation = data
+            } else {
+                throw ValidationError.invalidHexData
+            }
+        }
+
         public var description: String {
             "\(self.algorithm)/\(self.dataRepresentation.map { String(format: "%02hhx", $0) }.joined())"
         }
@@ -30,7 +42,7 @@ extension Biscuit {
         public var debugDescription: String {
             self.description
         }
-    
+
         init(dataRepresentation: Data, algorithm: SigningAlgorithm) {
             self.dataRepresentation = dataRepresentation
             self.algorithm = algorithm
@@ -48,19 +60,24 @@ extension Biscuit {
             return proto
         }
 
-        func isValidExternalSignature(_ signature: Data, for block: Block, lastSig: Data, interner: BlockInternmentTable) throws {
+        func isValidExternalSignature(
+            _ signature: Data,
+            for block: Block,
+            lastSig: Data,
+            interner: BlockInternmentTable
+        ) throws {
             let input = try SignatureV1.externalSignatureInput(block: block.datalog, sig: lastSig, interner: interner)
             switch self.algorithm.wrapped {
-                case .ed25519: 
-                    let key = try Curve25519.Signing.PublicKey(rawRepresentation: self.dataRepresentation)
-                    guard key.isValidSignature(signature, for: input) else {
-                        throw ValidationError.invalidExternalSignature
-                    }
-                case .secp256r1:
-                    let key = try P256.Signing.PublicKey(compressedRepresentation: self.dataRepresentation)
-                    guard key.isValidSignature(signature, for: input) else {
-                        throw ValidationError.invalidExternalSignature
-                    }
+            case .ed25519:
+                let key = try Curve25519.Signing.PublicKey(rawRepresentation: self.dataRepresentation)
+                guard key.isValidSignature(signature, for: input) else {
+                    throw ValidationError.invalidExternalSignature
+                }
+            case .secp256r1:
+                let key = try P256.Signing.PublicKey(compressedRepresentation: self.dataRepresentation)
+                guard key.isValidSignature(signature, for: input) else {
+                    throw ValidationError.invalidExternalSignature
+                }
             }
         }
     }
