@@ -12,7 +12,7 @@ import Foundation
 public struct Expression: ExpressionConvertible, Sendable, Hashable, CustomStringConvertible {
     let ops: [Op]
 
-    init(proto: Biscuit_Format_Schema_ExpressionV2, interner: BlockInternmentTable) throws {
+    init(proto: Biscuit_Format_Schema_Expression, interner: BlockInternmentTable) throws {
         self.ops = try proto.ops.map { try Op(proto: $0, interner: interner) }
     }
 
@@ -201,9 +201,9 @@ public struct Expression: ExpressionConvertible, Sendable, Hashable, CustomStrin
     public func tryOr<Rhs: ExpressionConvertible>(_ rhs: Rhs) -> Expression {
         Expression(op: .tryOr, lhs: Closure(body: self), rhs: rhs.expression)
     }
-    
-    func proto(_ interner: BlockInternmentTable) -> Biscuit_Format_Schema_ExpressionV2 {
-        var proto = Biscuit_Format_Schema_ExpressionV2()
+
+    func proto(_ interner: BlockInternmentTable) -> Biscuit_Format_Schema_Expression {
+        var proto = Biscuit_Format_Schema_Expression()
         proto.ops = self.ops.map { $0.proto(interner) }
         return proto
     }
@@ -212,23 +212,23 @@ public struct Expression: ExpressionConvertible, Sendable, Hashable, CustomStrin
         var stack: [StackElement] = []
         for op in self.ops {
             switch op {
-                case .value(let term):
-                    try stack.append(StackElement(term, variables))
-                case .unary(let op):
-                    guard let arg = stack.popLast() else {
-                        throw Biscuit.EvaluationError.invalidUnaryOp
-                    }
-                    try stack.append(.value(op.evaluate(arg)))
-                case .binary(let op):
-                    guard let arg2 = stack.popLast() else {
-                        throw Biscuit.EvaluationError.invalidBinaryOp
-                    }
-                    guard let arg1 = stack.popLast() else {
-                        throw Biscuit.EvaluationError.invalidBinaryOp
-                    }
-                    try stack.append(.value(op.evaluate(arg1, arg2, variables)))
-                case .closure(let closure):
-                    stack.append(.closure(closure))
+            case .value(let term):
+                try stack.append(StackElement(term, variables))
+            case .unary(let op):
+                guard let arg = stack.popLast() else {
+                    throw Biscuit.EvaluationError.invalidUnaryOp
+                }
+                try stack.append(.value(op.evaluate(arg)))
+            case .binary(let op):
+                guard let arg2 = stack.popLast() else {
+                    throw Biscuit.EvaluationError.invalidBinaryOp
+                }
+                guard let arg1 = stack.popLast() else {
+                    throw Biscuit.EvaluationError.invalidBinaryOp
+                }
+                try stack.append(.value(op.evaluate(arg1, arg2, variables)))
+            case .closure(let closure):
+                stack.append(.closure(closure))
             }
         }
         if stack.count == 1 {
@@ -264,30 +264,31 @@ enum Op: Sendable, Hashable {
     case closure(Closure)
 
     init(proto: Biscuit_Format_Schema_Op, interner: BlockInternmentTable) throws {
-        self = switch proto.content {
+        self =
+            switch proto.content {
             case .value(let term): try .value(Term(proto: term, interner: interner))
             case .unary(let op_unary): try .unary(OpUnary(proto: op_unary, interner: interner))
             case .binary(let op_binary): try .binary(OpBinary(proto: op_binary, interner: interner))
             case .closure(let op_closure): try .closure(Closure(proto: op_closure, interner: interner))
             case .none: throw Biscuit.ValidationError.missingOp
-        }
+            }
     }
 
     func intern(_ interner: inout BlockInternmentTable, _ locals: inout [String]) {
         switch self {
-            case .value(let term): term.intern(&interner, &locals)
-            case .closure(let closure): closure.intern(&interner, &locals)
-            default: return
+        case .value(let term): term.intern(&interner, &locals)
+        case .closure(let closure): closure.intern(&interner, &locals)
+        default: return
         }
     }
 
     func proto(_ interner: BlockInternmentTable) -> Biscuit_Format_Schema_Op {
         var proto = Biscuit_Format_Schema_Op()
         switch self {
-            case .value(let term): proto.value = term.proto(interner)
-            case .unary(let op_unary): proto.unary = op_unary.proto(interner)
-            case .binary(let op_binary): proto.binary = op_binary.proto(interner)
-            case .closure(let op_closure): proto.closure = op_closure.proto(interner)
+        case .value(let term): proto.value = term.proto(interner)
+        case .unary(let op_unary): proto.unary = op_unary.proto(interner)
+        case .binary(let op_binary): proto.binary = op_binary.proto(interner)
+        case .closure(let op_closure): proto.closure = op_closure.proto(interner)
         }
         return proto
     }
@@ -299,12 +300,12 @@ enum StackElement: Hashable {
 
     init(_ term: Term, _ variables: [String: Value]) throws {
         switch term.wrapped {
-            case .variable(let name):
-                guard let concrete = variables[name] else {
-                    throw Biscuit.EvaluationError.unknownVariable
-                }
-                self = .value(concrete)
-            case .value(let value): self = .value(value)
+        case .variable(let name):
+            guard let concrete = variables[name] else {
+                throw Biscuit.EvaluationError.unknownVariable
+            }
+            self = .value(concrete)
+        case .value(let value): self = .value(value)
         }
     }
 }
@@ -316,66 +317,66 @@ public protocol ExpressionConvertible {
 
 func parseIntoOps(_ parse: Parser.Expression, _ ops: inout [Op]) {
     switch parse {
-        case .term(let term): ops.append(.value(term))
-        case .closure(let closure): ops.append(.closure(closure))
-        case .unaryOp(let op, let e):
-            parseIntoOps(e, &ops)
-            ops.append(.unary(op))
-        case .binaryOp(let op, let e1, let e2):
-            parseIntoOps(e1, &ops)
-            parseIntoOps(e2, &ops)
-            ops.append(.binary(op))
+    case .term(let term): ops.append(.value(term))
+    case .closure(let closure): ops.append(.closure(closure))
+    case .unaryOp(let op, let e):
+        parseIntoOps(e, &ops)
+        ops.append(.unary(op))
+    case .binaryOp(let op, let e1, let e2):
+        parseIntoOps(e1, &ops)
+        parseIntoOps(e2, &ops)
+        ops.append(.binary(op))
     }
 }
 
 func printOps(_ ops: [Op], _ offset: inout Int, _ isLazy: Bool = false) -> String {
     offset -= 1
     switch ops[offset] {
-        case .value(let term): return term.description
-        case .closure(let closure): return isLazy ? closure.bodyDescription : closure.description
-        case .binary(let op):
-            let e2 = printOps(ops, &offset, op == .lazyAnd || op == .lazyOr)
-            let e1 = printOps(ops, &offset, op == .tryOr)
-            return switch op {
-                case .gt: "\(e1) > \(e2)"
-                case .gtEq: "\(e1) >= \(e2)"
-                case .lt: "\(e1) < \(e2)"
-                case .ltEq: "\(e1) <= \(e2)"
-                case .eq: "\(e1) === \(e2)"
-                case .notEq: "\(e1) !== \(e2)"
-                case .and: "\(e1) && \(e2)"
-                case .or: "\(e1) || \(e2)"
-                case .add: "\(e1) + \(e2)"
-                case .sub: "\(e1) - \(e2)"
-                case .mul: "\(e1) * \(e2)"
-                case .div: "\(e1) / \(e2)"
-                case .bitwiseAnd: "\(e1) & \(e2)"
-                case .bitwiseOr: "\(e1) | \(e2)"
-                case .bitwiseXor: "\(e1) ^ \(e2)"
-                case .contains: "\(e1).contains(\(e2))"
-                case .startsWith: "\(e1).starts_with(\(e2))"
-                case .endsWith: "\(e1).ends_with(\(e2))"
-                case .regex: "\(e1).matches(\(e2))"
-                case .intersection: "\(e1).intersection(\(e2))"
-                case .union: "\(e1).union(\(e2))"
-                case .heterogeneousEqual: "\(e1) == \(e2)"
-                case .heterogeneousNotEqual: "\(e1) != \(e2)"
-                case .lazyAnd: "\(e1) && \(e2)"
-                case .lazyOr: "\(e1) || \(e2)"
-                case .any: "\(e1).any(\(e2))"
-                case .all: "\(e1).all(\(e2))"
-                case .get: "\(e1).get(\(e2))"
-                case .ffi(let name): "\(e1).extern::\(name)(\(e2))"
-                case .tryOr: "\(e1).tryOr(\(e2))"
-            }
-        case .unary(let op):
-            let e = printOps(ops, &offset)
-            return switch op {
-                case .negate: "!\(e)"
-                case .length: "\(e).length()"
-                case .typeOf: "\(e).type()"
-                case .ffi(let name): "\(e).extern::\(name)()"
-                case .parens: "(\(e))"
-            }
+    case .value(let term): return term.description
+    case .closure(let closure): return isLazy ? closure.bodyDescription : closure.description
+    case .binary(let op):
+        let e2 = printOps(ops, &offset, op == .lazyAnd || op == .lazyOr)
+        let e1 = printOps(ops, &offset, op == .tryOr)
+        return switch op {
+        case .gt: "\(e1) > \(e2)"
+        case .gtEq: "\(e1) >= \(e2)"
+        case .lt: "\(e1) < \(e2)"
+        case .ltEq: "\(e1) <= \(e2)"
+        case .eq: "\(e1) === \(e2)"
+        case .notEq: "\(e1) !== \(e2)"
+        case .and: "\(e1) && \(e2)"
+        case .or: "\(e1) || \(e2)"
+        case .add: "\(e1) + \(e2)"
+        case .sub: "\(e1) - \(e2)"
+        case .mul: "\(e1) * \(e2)"
+        case .div: "\(e1) / \(e2)"
+        case .bitwiseAnd: "\(e1) & \(e2)"
+        case .bitwiseOr: "\(e1) | \(e2)"
+        case .bitwiseXor: "\(e1) ^ \(e2)"
+        case .contains: "\(e1).contains(\(e2))"
+        case .startsWith: "\(e1).starts_with(\(e2))"
+        case .endsWith: "\(e1).ends_with(\(e2))"
+        case .regex: "\(e1).matches(\(e2))"
+        case .intersection: "\(e1).intersection(\(e2))"
+        case .union: "\(e1).union(\(e2))"
+        case .heterogeneousEqual: "\(e1) == \(e2)"
+        case .heterogeneousNotEqual: "\(e1) != \(e2)"
+        case .lazyAnd: "\(e1) && \(e2)"
+        case .lazyOr: "\(e1) || \(e2)"
+        case .any: "\(e1).any(\(e2))"
+        case .all: "\(e1).all(\(e2))"
+        case .get: "\(e1).get(\(e2))"
+        case .ffi(let name): "\(e1).extern::\(name)(\(e2))"
+        case .tryOr: "\(e1).tryOr(\(e2))"
+        }
+    case .unary(let op):
+        let e = printOps(ops, &offset)
+        return switch op {
+        case .negate: "!\(e)"
+        case .length: "\(e).length()"
+        case .typeOf: "\(e).type()"
+        case .ffi(let name): "\(e).extern::\(name)()"
+        case .parens: "(\(e))"
+        }
     }
 }
