@@ -5,6 +5,7 @@
 import Crypto
 import SwiftProtobuf
 import XCTest
+
 @testable import Biscuits
 
 final class AuthorizationTests: XCTestCase {
@@ -25,12 +26,17 @@ final class AuthorizationTests: XCTestCase {
     func testBasicToken() throws {
         let biscuit = try self.loadBiscuit(from: "test001_basic")
         do {
-            try biscuit.authorize(using: """
-                resource("file1");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file1");
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
-            try XCTAssertEqual(err.failedCheck, Check("check if resource($0), operation(\"read\"), right($0, \"read\");"))
+            try XCTAssertEqual(
+                err.failedCheck,
+                Check("check if resource($0), operation(\"read\"), right($0, \"read\");")
+            )
             return
         }
         XCTAssert(false)
@@ -39,16 +45,19 @@ final class AuthorizationTests: XCTestCase {
     func testBasicTokenDSL() throws {
         let biscuit = try self.loadBiscuit(from: "test001_basic")
         do {
-            try biscuit.authorize() {
+            try biscuit.authorize {
                 Fact("resource", "file1")
                 Policy.alwaysAllow
             }
         } catch let err as Biscuit.AuthorizationError {
-            XCTAssertEqual(err.failedCheck, Check.checkIf {
-                Predicate("resource", Term(variable: "0"))
-                Predicate("operation", "read")
-                Predicate("right", Term(variable: "0"), "read")
-            })
+            XCTAssertEqual(
+                err.failedCheck,
+                Check.checkIf {
+                    Predicate("resource", Term(variable: "0"))
+                    Predicate("operation", "read")
+                    Predicate("right", Term(variable: "0"), "read")
+                }
+            )
             return
         }
         XCTAssert(false)
@@ -102,13 +111,18 @@ final class AuthorizationTests: XCTestCase {
     func testScopedRules() throws {
         let biscuit = try self.loadBiscuit(from: "test007_scoped_rules")
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                operation("read");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        operation("read");
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
-            try XCTAssertEqual(err.failedCheck, Check("check if resource($0), operation(\"read\"), right($0, \"read\");"))
+            try XCTAssertEqual(
+                err.failedCheck,
+                Check("check if resource($0), operation(\"read\"), right($0, \"read\");")
+            )
             return
         }
         XCTAssert(false)
@@ -117,13 +131,18 @@ final class AuthorizationTests: XCTestCase {
     func testScopedChecks() throws {
         let biscuit = try self.loadBiscuit(from: "test008_scoped_checks")
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                operation("read");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        operation("read");
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
-            try XCTAssertEqual(err.failedCheck, Check("check if resource($0), operation(\"read\"), right($0, \"read\");"))
+            try XCTAssertEqual(
+                err.failedCheck,
+                Check("check if resource($0), operation(\"read\"), right($0, \"read\");")
+            )
             return
         }
         XCTAssert(false)
@@ -132,12 +151,14 @@ final class AuthorizationTests: XCTestCase {
     func testExpiredToken() throws {
         let biscuit = try self.loadBiscuit(from: "test009_expired_token")
         do {
-            try biscuit.authorize(using: """
-                resource("file1");
-                operation("read");
-                time(2020-12-21T09:23:12Z);
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file1");
+                        operation("read");
+                        time(2020-12-21T09:23:12Z);
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
             try XCTAssertEqual(err.failedCheck, Check("check if time($time), $time <= 2018-12-20T00:00:00Z;"))
             return
@@ -148,12 +169,14 @@ final class AuthorizationTests: XCTestCase {
     func testAuthorizerScope() throws {
         let biscuit = try self.loadBiscuit(from: "test010_authorizer_scope")
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                operation("read");
-                check if right($0, $1), resource($0), operation($1);
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        operation("read");
+                        check if right($0, $1), resource($0), operation($1);
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
             try XCTAssertEqual(err.failedCheck, Check("check if right($0, $1), resource($0), operation($1);"))
             return
@@ -164,12 +187,14 @@ final class AuthorizationTests: XCTestCase {
     func testAuthorizerAuthorityChecks() throws {
         let biscuit = try self.loadBiscuit(from: "test011_authorizer_authority_caveats")
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                operation("read");
-                check if right($0, $1), resource($0), operation($1);
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        operation("read");
+                        check if right($0, $1), resource($0), operation($1);
+                        allow if true;
+                    """
+            )
         } catch let err as Biscuit.AuthorizationError {
             try XCTAssertEqual(err.failedCheck, Check("check if right($0, $1), resource($0), operation($1);"))
             return
@@ -180,11 +205,13 @@ final class AuthorizationTests: XCTestCase {
     func testAuthorityChecks() throws {
         let biscuit = try self.loadBiscuit(from: "test012_authority_caveats")
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                operation("read");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        operation("read");
+                        allow if true;
+                    """
+            )
         } catch _ as Biscuit.AuthorizationError {
             return
         }
@@ -193,17 +220,21 @@ final class AuthorizationTests: XCTestCase {
 
     func testBlockRules() throws {
         let biscuit = try self.loadBiscuit(from: "test013_block_rules")
-        try biscuit.authorize(using: """
-            resource("file1");
-            time(2020-12-21T09:23:12Z);
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    resource("file1");
+                    time(2020-12-21T09:23:12Z);
+                    allow if true;
+                """
+        )
         do {
-            try biscuit.authorize(using: """
-                resource("file2");
-                time(2020-12-21T09:23:12Z);
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file2");
+                        time(2020-12-21T09:23:12Z);
+                        allow if true;
+                    """
+            )
         } catch _ as Biscuit.AuthorizationError {
             return
         }
@@ -213,15 +244,19 @@ final class AuthorizationTests: XCTestCase {
 
     func testRegexConstraint() throws {
         let biscuit = try self.loadBiscuit(from: "test014_regex_constraint")
-        try biscuit.authorize(using: """
-            resource("file123.txt");
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    resource("file123.txt");
+                    allow if true;
+                """
+        )
         do {
-            try biscuit.authorize(using: """
-                resource("file1");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        resource("file1");
+                        allow if true;
+                    """
+            )
         } catch _ as Biscuit.AuthorizationError {
             return
         }
@@ -230,10 +265,12 @@ final class AuthorizationTests: XCTestCase {
 
     func testMultiQueriesCaveats() throws {
         let biscuit = try self.loadBiscuit(from: "test015_multi_queries_caveats")
-        try biscuit.authorize(using: """
-            check if must_be_present($0) or must_be_present($0);
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    check if must_be_present($0) or must_be_present($0);
+                    allow if true;
+                """
+        )
     }
 
     func testIndependentCheckHeadName() throws {
@@ -273,31 +310,37 @@ final class AuthorizationTests: XCTestCase {
 
     func testSealed() throws {
         let biscuit = try self.loadBiscuit(from: "test020_sealed")
-        try biscuit.authorize(using: """
-            resource("file1");
-            operation("read");
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    resource("file1");
+                    operation("read");
+                    allow if true;
+                """
+        )
     }
 
     func testParsing() throws {
         let biscuit = try self.loadBiscuit(from: "test021_parsing")
-        try biscuit.authorize(using: """
-            check if ns::fact_123("hello é\t😁");
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    check if ns::fact_123("hello é\t😁");
+                    allow if true;
+                """
+        )
     }
 
     func testDefaultSymbols() throws {
         let biscuit = try self.loadBiscuit(from: "test022_default_symbols")
-        try biscuit.authorize(using: """
-            check if read(0), write(1), resource(2), operation(3), right(4), time(5),
-                     role(6), owner(7), tenant(8), namespace(9), user(10), team(11),
-                     service(12), admin(13), email(14), group(15), member(16), ip_address(17),
-                     client(18), client_ip(19), domain(20), path(21), version(22), cluster(23),
-                     node(24), hostname(25), nonce(26), query(27);
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    check if read(0), write(1), resource(2), operation(3), right(4), time(5),
+                             role(6), owner(7), tenant(8), namespace(9), user(10), team(11),
+                             service(12), admin(13), email(14), group(15), member(16), ip_address(17),
+                             client(18), client_ip(19), domain(20), path(21), version(22), cluster(23),
+                             node(24), hostname(25), nonce(26), query(27);
+                    allow if true;
+                """
+        )
     }
 
     func testExecutionScope() throws {
@@ -309,7 +352,7 @@ final class AuthorizationTests: XCTestCase {
         }
         XCTAssert(false)
     }
-    
+
     func testThirdParty() throws {
         let biscuit = try self.loadBiscuit(from: "test024_third_party")
         try biscuit.authorize(using: "allow if true;")
@@ -317,17 +360,21 @@ final class AuthorizationTests: XCTestCase {
 
     func testCheckAll() throws {
         let biscuit = try self.loadBiscuit(from: "test025_check_all")
-        try biscuit.authorize(using: """
-            operation("A");
-            operation("B");
-            allow if true;
-        """)
+        try biscuit.authorize(
+            using: """
+                    operation("A");
+                    operation("B");
+                    allow if true;
+                """
+        )
         do {
-            try biscuit.authorize(using: """
-                operation("A");
-                operation("invalid");
-                allow if true;
-            """)
+            try biscuit.authorize(
+                using: """
+                        operation("A");
+                        operation("invalid");
+                        allow if true;
+                    """
+            )
         } catch _ as Biscuit.AuthorizationError {
             return
         }
@@ -336,14 +383,16 @@ final class AuthorizationTests: XCTestCase {
 
     func testPublicKeysInterning() throws {
         let biscuit = try self.loadBiscuit(from: "test026_public_keys_interning")
-        try biscuit.authorize(using: """
-            check if query(1, 2) trusting ed25519/acdd6d5b53bfee478bf689f8e012fe7988bf755e3d7c5152947abc149bc20189, ed25519/a060270db7e9c9f06e8f9cc33a64e99f6596af12cb01c4b638df8afc7b642463;
+        try biscuit.authorize(
+            using: """
+                    check if query(1, 2) trusting ed25519/acdd6d5b53bfee478bf689f8e012fe7988bf755e3d7c5152947abc149bc20189, ed25519/a060270db7e9c9f06e8f9cc33a64e99f6596af12cb01c4b638df8afc7b642463;
 
-            deny if query(3);
-            deny if query(1, 2);
-            deny if query(0) trusting ed25519/acdd6d5b53bfee478bf689f8e012fe7988bf755e3d7c5152947abc149bc20189;
-            allow if true;
-        """)
+                    deny if query(3);
+                    deny if query(1, 2);
+                    deny if query(0) trusting ed25519/acdd6d5b53bfee478bf689f8e012fe7988bf755e3d7c5152947abc149bc20189;
+                    allow if true;
+                """
+        )
     }
 
     func testIntegerWraparound() throws {
@@ -399,7 +448,7 @@ final class AuthorizationTests: XCTestCase {
 
     func testLazinessClosures() throws {
         let biscuit = try self.loadBiscuit(from: "test032_laziness_closures")
-        try biscuit.authorize() { Policy.alwaysAllow }
+        try biscuit.authorize { Policy.alwaysAllow }
         do {
             try biscuit.authorize(using: "allow if [true].any($p -> [true].all($p -> $p));")
         } catch let err as Biscuit.EvaluationError where err == Biscuit.EvaluationError.variableShadowing {
@@ -410,12 +459,12 @@ final class AuthorizationTests: XCTestCase {
 
     func testTypeOf() throws {
         let biscuit = try self.loadBiscuit(from: "test033_typeof")
-        try biscuit.authorize() { Policy.alwaysAllow }
+        try biscuit.authorize { Policy.alwaysAllow }
     }
 
     func testArrayMap() throws {
         let biscuit = try self.loadBiscuit(from: "test034_array_map")
-        try biscuit.authorize() { Policy.alwaysAllow }
+        try biscuit.authorize { Policy.alwaysAllow }
     }
 
     func testFfiUnspecified() throws {
@@ -431,7 +480,7 @@ final class AuthorizationTests: XCTestCase {
 
     func testSecP256R1() throws {
         let biscuit = try self.loadBiscuit(from: "test036_secp256r1")
-        try biscuit.authorize() {
+        try biscuit.authorize {
             Fact("resource", "file1")
             Fact("operation", "read")
             Policy.alwaysAllow
@@ -440,7 +489,7 @@ final class AuthorizationTests: XCTestCase {
 
     func testSecP256R1ThirdParty() throws {
         let biscuit = try self.loadBiscuit(from: "test037_secp256r1_third_party")
-        try biscuit.authorize() {
+        try biscuit.authorize {
             Fact("resource", "file1")
             Fact("operation", "read")
             Policy.alwaysAllow
@@ -449,7 +498,7 @@ final class AuthorizationTests: XCTestCase {
 
     func testTryOp() throws {
         let biscuit = try self.loadBiscuit(from: "test038_try_op")
-        try biscuit.authorize() { Policy.alwaysAllow }
+        try biscuit.authorize { Policy.alwaysAllow }
         do {
             try biscuit.authorize(using: "check if true.try_or(true === 12); allow if true;")
         } catch let err as Biscuit.EvaluationError where err == Biscuit.EvaluationError.typeError {
@@ -482,11 +531,11 @@ func htoi(_ value: UInt8) -> UInt8? {
     let charA = UInt8(UnicodeScalar("a").value)
 
     switch value {
-        case char0...char0 + 9:
-            return value - char0
-        case charA...charA + 5:
-            return value - charA + 10
-        default:
-            return nil
+    case char0...char0 + 9:
+        return value - char0
+    case charA...charA + 5:
+        return value - charA + 10
+    default:
+        return nil
     }
 }
